# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Onyx Domain API V1",
    "version" : "1.0.9"
  },
  "servers" : [ {
    "url" : "http://onyxdomainapi.azurewebsites.net"
  } ],
  "paths" : {
...
    "/Accounts/Save" : {
      "post" : {
        "tags" : [ "Accounts" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AccountData"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AccountData"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AccountData"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
    "/Accounts/Delete" : {
      "post" : {
        "tags" : [ "Accounts" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IdOnly"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IdOnly"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IdOnly"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
...
    "/Brokers/Save" : {
      "post" : {
        "tags" : [ "Brokers" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BrokerData"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BrokerData"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BrokerData"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
    "/Brokers/Delete" : {
      "post" : {
        "tags" : [ "Brokers" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IdOnly"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IdOnly"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IdOnly"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
    "/CapitalActivities/List" : {
      "get" : {
        "tags" : [ "CapitalActivities" ],
        "parameters" : [ {
          "name" : "f",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "nullable" : true
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
    "/CapitalActivities/Detail" : {
      "get" : {
        "tags" : [ "CapitalActivities" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "nullable" : true
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
    "/CapitalActivities/Delete" : {
      "post" : {
        "tags" : [ "CapitalActivities" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IdOnly"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IdOnly"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IdOnly"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
    "/CapitalActivities/Save" : {
      "post" : {
        "tags" : [ "CapitalActivities" ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CapitalActivityData"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CapitalActivityData"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CapitalActivityData"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              },
              "text/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Context"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Context" : {
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string",
            "nullable" : true
          },
          "success" : {
            "type" : "boolean"
          },
          "data" : {
            "nullable" : true
          }
        },
        "additionalProperties" : false,
        "example" : {
          "data" : "",
          "success" : true,
          "message" : "message"
        }
      },
...
      "AccountData" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "name" : {
            "type" : "string",
            "nullable" : true
          },
          "currency" : {
            "type" : "string",
            "nullable" : true
          },
          "currencyId" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "additionalProperties" : false
      },
      "IdOnly" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "additionalProperties" : false
      },
...
      "IdN" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "n" : {
            "type" : "string",
            "nullable" : true
          }
        },
        "additionalProperties" : false
      },
      "BrokerData" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "name" : {
            "type" : "string",
            "nullable" : true
          },
          "code" : {
            "type" : "string",
            "nullable" : true
          },
          "brokerageType" : {
            "$ref" : "#/components/schemas/IdN"
          },
          "stockExchange" : {
            "$ref" : "#/components/schemas/IdN"
          },
          "brokerage" : {
            "type" : "number",
            "format" : "double"
          },
          "significantDigits" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "additionalProperties" : false
      },
      "CapitalActivityData" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "investimentCode" : {
            "type" : "integer",
            "format" : "int32"
          },
          "date" : {
            "type" : "string",
            "format" : "date-time",
            "nullable" : true
          },
          "amount" : {
            "type" : "number",
            "format" : "double"
          },
          "factor" : {
            "type" : "number",
            "format" : "double"
          },
          "shareClassId" : {
            "type" : "integer",
            "format" : "int32",
            "nullable" : true
          },
          "fundId" : {
            "type" : "integer",
            "format" : "int32",
            "nullable" : true
          }
        },
        "additionalProperties" : false
      },
...
    }
  }
}